Started on the map
Started on the map, laid out 3 buildings, with a U shaped road and sidewalks. (These are all subject to change I just wanted to get a rough outline going). Also added a skybox texture for the night sky (you may find it easier to change this to default while developing). I also added some of the assets we are going to use like the fog, although haven't implemented it yet.