Added first person
Added first person camera and character controller. The player can walk and run with varying speeds that we can adjust as needed. Hold shift to run. I also added sounds for footsteps for walking and running, with intervals that can also be changed. We can also adjust these footstep sounds if we feel like we should but they are good for now.